feat: Add depth endpoint for category API
This commit adds a new file `depth.js` in the `category/controllers` directory, which contains an asynchronous function `depth` that retrieves data from the `api::category.depth` service and sets it as the response body.

It also adds a new file `depth.js` in the `category/routes` directory, which defines a route for the GET method at `/categories_depth`. The handler for this route is set to `depth.depth`.

In addition, a new file `depth.js` is added in the `category/services` directory. This file exports an asynchronous function named `depth`, which retrieves existing categories using the `strapi.entityService.findMany("api::category.category")`. If no categories are found, it throws an error.

Lastly, there is a modification in the full_documentation.json file where the "x-generation-date" field is updated to "2023-09-16T11:34:12.630Z".

These changes enable retrieving category depth information through the newly added endpoint.