Update x-generation-date in full_documentation.json
The commit updates the x-generation-date field in the full_documentation.json file. The new value is "2023-09-16T13:44:48.932Z".

Add order and depth fields to multiple sections

This commit adds the order and depth fields to several sections in the full_documentation.json file. These fields are of type integer.

Update createdAt field in multiple sections

The commit updates the createdAt field in multiple sections of the full_documentation.json file. The field is of type string with format "date-time".