🔀 Refactor allCompany service to remove unnecessary properties recursively
This commit refactors the `allCompany` service in order to remove unnecessary properties recursively from the returned data. The code now uses a helper function, `removePropsRecursively`, to delete the `createdBy` and `updatedBy` properties from each object in the data array. This ensures that sensitive information is not exposed when retrieving company records.